major refactoring of all source space functions that are more general freesurfer functions
