+++ mne/utils/__init__.py
@@ -10,7 +10,7 @@ from .check import (check_fname, check_version, check_random_state,
                     _check_fname, _check_subject, _check_pandas_installed,
                     _check_pandas_index_arguments, _check_mayavi_version,
                     _check_event_id, _check_ch_locs, _check_compensation_grade,
-                    _check_if_nan, _is_numeric, _ensure_int,  _check_preload,
+                    _check_if_nan, _is_numeric, _ensure_int, _check_preload,
                     _validate_type, _check_info_inv, _check_pylsl_installed,
                     _check_channels_spatial_filter, _check_one_ch_type,
                     _check_rank, _check_option, _check_depth, _check_combine,
@@ -21,9 +21,9 @@ from .check import (check_fname, check_version, check_random_state,
                     _check_all_same_channel_names, path_like, _ensure_events,
                     _check_eeglabio_installed)
 from .config import (set_config, get_config, get_config_path, set_cache_dir,
-                     set_memmap_min_size, get_subjects_dir, _get_stim_channel,
+                     set_memmap_min_size, _get_stim_channel,
                      sys_info, _get_extra_data_path, _get_root_dir,
-                     _get_numpy_libs, read_freesurfer_lut)
+                     _get_numpy_libs)
 from .docs import (copy_function_doc_to_method_doc, copy_doc, linkcode_resolve,
                    open_docs, deprecated, fill_doc, deprecated_alias,
                    copy_base_doc_to_subclass_doc, docdict as _docdict)
@@ -45,10 +45,11 @@ from ._testing import (run_command_if_main, requires_sklearn,
                        _TempDir, has_nibabel, _import_mlab, buggy_mkl_svd,
                        requires_numpydoc, requires_vtk, requires_freesurfer,
                        requires_nitime, requires_dipy,
-                       requires_neuromag2ft, requires_pylsl, assert_object_equal,
-                       assert_and_remove_boundary_annot, _raw_annot,
-                       assert_dig_allclose, assert_meg_snr, assert_snr,
-                       assert_stcs_equal, modified_env, _click_ch_name)
+                       requires_neuromag2ft, requires_pylsl,
+                       assert_object_equal, assert_and_remove_boundary_annot,
+                       _raw_annot, assert_dig_allclose, assert_meg_snr,
+                       assert_snr, assert_stcs_equal, modified_env,
+                       _click_ch_name)
 from .numerics import (hashfunc, _compute_row_norms,
                        _reg_pinv, random_permutation, _reject_data_segments,
                        compute_corr, _get_inst_data, array_split_idx,
